backend/requirements.txt
Pygments==2.14.0
PyYAML==6.0
six==1.16.0

pydantic~=2.5.2
fastapi~=0.104.1
python-dotenv~=1.0.0
requests~=2.31.0
langchain~=0.0.340
openai~=1.3.5
chromadb~=0.3.29
